remove AbstractCuratorServerInventoryView
